Moved the EntityComponentID type out of the SpatialRPCService.
